Refactor code to apply ruby best practices
